5225: Move helpers into flow appropriate packages
To make it easier to navigate through the codebase we've split the
helpers package into four bits that represent the four legs of the user
journey through the bridge: requestFromVerify, requestToEidas,
responseFromEidas and responseToVerify.

Authors: @richardtowers @sfkamath
