+++ src/main/java/uk/gov/ida/eidas/bridge/factories/VerifyEidasBridgeFactory.java
@@ -17,20 +17,20 @@ import org.opensaml.xmlsec.keyinfo.impl.X509KeyInfoGeneratorFactory;
 import org.opensaml.xmlsec.signature.support.impl.ExplicitKeySignatureTrustEngine;
 import uk.gov.ida.eidas.bridge.configuration.BridgeConfiguration;
 import uk.gov.ida.eidas.bridge.configuration.KeyStoreConfiguration;
-import uk.gov.ida.eidas.bridge.helpers.AssertionConsumerServiceLocator;
-import uk.gov.ida.eidas.bridge.helpers.AssertionSubjectGenerator;
-import uk.gov.ida.eidas.bridge.helpers.AuthnRequestFormGenerator;
-import uk.gov.ida.eidas.bridge.helpers.AuthnRequestHandler;
-import uk.gov.ida.eidas.bridge.helpers.AuthnStatementAssertionGenerator;
-import uk.gov.ida.eidas.bridge.helpers.EidasAuthnRequestGenerator;
-import uk.gov.ida.eidas.bridge.helpers.EidasIdentityAssertionUnmarshaller;
-import uk.gov.ida.eidas.bridge.helpers.MatchingDatasetAssertionGenerator;
-import uk.gov.ida.eidas.bridge.helpers.MetadataBackedEncryptionPublicKeyRetriever;
-import uk.gov.ida.eidas.bridge.helpers.ResponseHandler;
-import uk.gov.ida.eidas.bridge.helpers.ResponseSizeValidator;
+import uk.gov.ida.eidas.bridge.helpers.responseToVerify.AssertionConsumerServiceLocator;
+import uk.gov.ida.eidas.bridge.helpers.responseToVerify.AssertionSubjectGenerator;
+import uk.gov.ida.eidas.bridge.helpers.requestToEidas.AuthnRequestFormGenerator;
+import uk.gov.ida.eidas.bridge.helpers.requestFromVerify.AuthnRequestHandler;
+import uk.gov.ida.eidas.bridge.helpers.responseToVerify.AuthnStatementAssertionGenerator;
+import uk.gov.ida.eidas.bridge.helpers.requestToEidas.EidasAuthnRequestGenerator;
+import uk.gov.ida.eidas.bridge.helpers.responseFromEidas.EidasIdentityAssertionUnmarshaller;
+import uk.gov.ida.eidas.bridge.helpers.responseToVerify.MatchingDatasetAssertionGenerator;
+import uk.gov.ida.eidas.bridge.helpers.responseToVerify.MetadataBackedEncryptionPublicKeyRetriever;
+import uk.gov.ida.eidas.bridge.helpers.responseFromEidas.ResponseHandler;
+import uk.gov.ida.eidas.bridge.helpers.responseFromEidas.ResponseSizeValidator;
 import uk.gov.ida.eidas.bridge.helpers.SigningHelper;
-import uk.gov.ida.eidas.bridge.helpers.SingleSignOnServiceLocator;
-import uk.gov.ida.eidas.bridge.helpers.VerifyResponseGenerator;
+import uk.gov.ida.eidas.bridge.helpers.requestToEidas.SingleSignOnServiceLocator;
+import uk.gov.ida.eidas.bridge.helpers.responseToVerify.VerifyResponseGenerator;
 import uk.gov.ida.eidas.bridge.resources.BridgeMetadataResource;
 import uk.gov.ida.eidas.bridge.resources.EidasResponseResource;
 import uk.gov.ida.eidas.bridge.resources.VerifyAuthnRequestResource;
