+++ src/main/java/uk/gov/ida/eidas/bridge/helpers/requestToEidas/EidasAuthnRequestGenerator.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.eidas.bridge.helpers;
+package uk.gov.ida.eidas.bridge.helpers.requestToEidas;
 
 import org.joda.time.DateTime;
 import org.opensaml.core.xml.Namespace;
@@ -17,6 +17,7 @@ import org.opensaml.saml.saml2.core.StatusResponseType;
 import org.opensaml.saml.saml2.core.impl.ExtensionsBuilder;
 import org.opensaml.security.SecurityException;
 import org.opensaml.xmlsec.signature.support.SignatureException;
+import uk.gov.ida.eidas.bridge.helpers.SigningHelper;
 import uk.gov.ida.eidas.common.LevelOfAssurance;
 import uk.gov.ida.eidas.saml.extensions.NamespaceConstants;
 import uk.gov.ida.eidas.saml.extensions.RequestedAttribute;
