+++ src/main/java/uk/gov/ida/eidas/bridge/helpers/responseToVerify/MetadataBackedEncryptionPublicKeyRetriever.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.eidas.bridge.helpers;
+package uk.gov.ida.eidas.bridge.helpers.responseToVerify;
 
 import com.google.common.base.Throwables;
 import net.shibboleth.utilities.java.support.resolver.CriteriaSet;
