+++ src/test/java/uk/gov/ida/eidas/bridge/helpers/responseFromEidas/EidasIdentityAssertionUnmarshallerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.eidas.bridge.helpers;
+package uk.gov.ida.eidas.bridge.helpers.responseFromEidas;
 
 import io.dropwizard.testing.ResourceHelpers;
 import org.joda.time.DateTime;
@@ -6,6 +6,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.opensaml.saml.saml2.core.Assertion;
 import uk.gov.ida.eidas.bridge.domain.EidasIdentityAssertion;
+import uk.gov.ida.eidas.bridge.helpers.EidasSamlBootstrap;
 import uk.gov.ida.saml.core.domain.Gender;
 import uk.gov.ida.saml.deserializers.parser.SamlObjectParser;
 import uk.gov.ida.saml.security.validators.ValidatedAssertions;
